+++ R/stat-dens1d-labels.r
@@ -8,12 +8,36 @@
 #'   to \code{rownames(data)}, with a message.
 #'
 #' @details \code{stat_dens1d_labels()} is designed to work together with
-#'   statistics from package 'ggrepel'. To avoid text labels being plotted over
+#'   geometries from packages 'ggrepel'. To avoid text labels being plotted over
 #'   unlabelled points the corresponding rows in data need to be retained but
-#'   labels replaced with the empty character string, \code{""}. This makes
-#'   \code{\link{stat_dens1d_filter}} unsuitable for the task. Non-the-less
-#'   \code{stat_dens1d_labels()} could be useful in some other cases, as the
-#'   substitution character string can be set by the user.
+#'   labels replaced with the empty character string, \code{""}. Function
+#'   \code{\link{stat_dens1d_filter}} cannot be used with the repulsive geoms
+#'   from 'ggrepel' because it drops the observations.
+#'
+#'   \code{stat_dens1d_labels()} can be useful also in other situations, as the
+#'   substitution character string can be set by the user by passing an argument
+#'   to \code{label.fill}. If this argument is \code{NULL} the unselected rows
+#'   are filtered out.
+#'
+#'   The local density of observations along \emph{x} or \emph{y} is computed
+#'   with function \code{\link[stats]{density}} and used to select observations,
+#'   passing to the geom all the rows in its \code{data} input but with with the
+#'   text of labels replaced in those "not kept". The default is to select
+#'   observations in sparse regions of the plot, but the selection can be
+#'   inverted so that only observations in the densest regions are returned.
+#'   Specific observations can be protected from having the label replaced by
+#'   passing a suitable argument to \code{keep.these}. Logical and integer
+#'   vectors function as indexes to rows in \code{data}, while a character
+#'   vector is compared to values in the variable mapped to the \code{label}
+#'   aesthetic. A function passed as argument to keep.these will receive as
+#'   argument the values in the variable mapped to \code{label} and should
+#'   return a character, logical or numeric vector as described above.
+#'
+#'   How many labels are retained intact in addition to those in
+#'   \code{keep.these} is controlled with arguments passed to \code{keep.number}
+#'   and \code{keep.fraction}. \code{keep.number} sets the maximum number of
+#'   observations selected, whenever \code{keep.fraction} results in fewer
+#'   observations selected, it is obeyed.
 #'
 #' @param mapping The aesthetic mapping, usually constructed with
 #'   \code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. Only needs to be
@@ -29,6 +53,10 @@
 #' @param keep.sparse logical If \code{TRUE}, the default, observations from the
 #'   more sparse regions are retained, if \code{FALSE} those from the densest
 #'   regions.
+#' @param keep.these character vector, integer vector, logical vector or
+#'   function that takes the variable mapped to the \code{label} aesthetic as
+#'   first argument and returns a character vector or a logical vector. These
+#'   rows from \code{data} are selected irrespective of the local density.
 #' @param invert.selection logical If \code{TRUE}, the complement of the
 #'   selected rows are returned.
 #' @param bw numeric or character The smoothing bandwidth to be used. If
@@ -155,6 +183,7 @@ stat_dens1d_labels <-
            keep.fraction = 0.10,
            keep.number = Inf,
            keep.sparse = TRUE,
+           keep.these = FALSE,
            invert.selection = FALSE,
            bw = "SJ",
            kernel = "gaussian",
@@ -165,6 +194,17 @@ stat_dens1d_labels <-
            na.rm = TRUE,
            show.legend = FALSE,
            inherit.aes = TRUE) {
+
+    if (length(label.fill) > 1L) {
+      stop("Length for 'label.fill' is not 1: ", label.fill)
+    }
+    if (is.na(keep.fraction) || keep.fraction < 0 || keep.fraction > 1) {
+      stop("Out of range or missing value for 'keep.fraction': ", keep.fraction)
+    }
+    if (is.na(keep.number) || keep.number < 0) {
+      stop("Out of range or missing value for 'keep.number': ", keep.number)
+    }
+
     ggplot2::layer(
       stat = StatDens1dLabels, data = data, mapping = mapping, geom = geom,
       position = position, show.legend = show.legend, inherit.aes = inherit.aes,
@@ -172,6 +212,7 @@ stat_dens1d_labels <-
                     keep.fraction = keep.fraction,
                     keep.number = keep.number,
                     keep.sparse = keep.sparse,
+                    keep.these = keep.these,
                     invert.selection = invert.selection,
                     bw = bw,
                     adjust = adjust,
@@ -189,6 +230,7 @@ dens1d_labs_compute_fun <-
            keep.fraction,
            keep.number,
            keep.sparse,
+           keep.these,
            invert.selection,
            bw,
            kernel,
@@ -197,25 +239,16 @@ dens1d_labs_compute_fun <-
            orientation,
            label.fill) {
 
-    if (length(label.fill) != 1L) {
-      stop("Length for 'label.fill' is not 1: ", label.fill)
-    }
-
-    if (is.na(keep.fraction) || keep.fraction < 0 || keep.fraction > 1) {
-      stop("Out of range or missing value for 'keep.fraction': ", keep.fraction)
-    }
-    if (is.na(keep.number) || keep.number < 0) {
-      stop("Out of range or missing value for 'keep.number': ", keep.number)
-    }
-
     force(data)
+    if (!exists("label", data)) {
+#      message("Mapping 'rownames(data)' to missing 'label' aesthetic")
+      data[["label"]] <- rownames(data)
+    }
+
+    keep.these <- keep_these2logical(keep.these = keep.these, data = data)
+
     if (nrow(data) * keep.fraction > keep.number) {
       keep.fraction <- keep.number / nrow(data)
-    }
-
-    if (!exists("label", data)) {
-      message("Mapping missing 'label' aesthetic to 'rownames(data)'.")
-      data[["label"]] <- rownames(data)
     }
 
     if (keep.fraction == 1) {
@@ -239,7 +272,13 @@ dens1d_labs_compute_fun <-
       }
     }
 
-    if (is.function(label.fill)) {
+    if (is.null(label.fill)) {
+      if (invert.selection){
+        data <- data[!(keep | keep.these), ]
+      } else {
+        data <- data[keep | keep.these, ]
+      }
+    } else if (is.function(label.fill)) {
       if (invert.selection){
         data[["label"]] <- ifelse(!keep,
                                   data[["label"]],
@@ -260,11 +299,11 @@ dens1d_labs_compute_fun <-
         }
       }
       if (invert.selection){
-        data[["label"]] <- ifelse(!keep,
+        data[["label"]] <- ifelse(!(keep | keep.these),
                                   data[["label"]],
                                   label.fill)
       } else {
-        data[["label"]] <- ifelse(keep,
+        data[["label"]] <- ifelse(keep | keep.these,
                                   data[["label"]],
                                   label.fill)
       }
@@ -283,5 +322,5 @@ StatDens1dLabels <-
     ggplot2::Stat,
     compute_panel =
       dens1d_labs_compute_fun,
-    required_aes = "x|y" # c("x", "y")
+    required_aes = "x|y"
   )