+++ R/stat-dens2d-labels.r
@@ -8,12 +8,36 @@
 #'   to \code{rownames(data)}, with a message.
 #'
 #' @details \code{stat_dens2d_labels()} is designed to work together with
-#'   statistics from package 'ggrepel'. To avoid text labels being plotted over
+#'   geometries from packages 'ggrepel'. To avoid text labels being plotted over
 #'   unlabelled points the corresponding rows in data need to be retained but
-#'   labels replaced with the empty character string, \code{""}. This makes
-#'   \code{\link{stat_dens2d_filter}} unsuitable for the task. Non-the-less
-#'   \code{stat_dens2d_labels()} could be useful in some other cases, as the
-#'   substitution character string can be set by the user.
+#'   labels replaced with the empty character string, \code{""}. Function
+#'   \code{\link{stat_dens2d_filter}} cannot be used with the repulsive geoms
+#'   from 'ggrepel' because it drops the observations.
+#'
+#'   \code{stat_dens2d_labels()} can be useful also in other situations, as the
+#'   substitution character string can be set by the user by passing an argument
+#'   to \code{label.fill}. If this argument is \code{NULL} the unselected rows
+#'   are filtered out.
+#'
+#'   The local density of observations in 2D (\emph{x} and \emph{y}) is computed
+#'   with function \code{\link[MASS]{kde2d}} and used to select observations,
+#'   passing to the geom all the rows in its \code{data} input but with with the
+#'   text of labels replaced in those "not kept". The default is to select
+#'   observations in sparse regions of the plot, but the selection can be
+#'   inverted so that only observations in the densest regions are returned.
+#'   Specific observations can be protected from having the label replaced by
+#'   passing a suitable argument to \code{keep.these}. Logical and integer
+#'   vectors function as indexes to rows in \code{data}, while a character
+#'   vector is compared to values in the variable mapped to the \code{label}
+#'   aesthetic. A function passed as argument to keep.these will receive as
+#'   argument the values in the variable mapped to \code{label} and should
+#'   return a character, logical or numeric vector as described above.
+#'
+#'   How many labels are retained intact in addition to those in
+#'   \code{keep.these} is controlled with arguments passed to \code{keep.number}
+#'   and \code{keep.fraction}. \code{keep.number} sets the maximum number of
+#'   observations selected, whenever \code{keep.fraction} results in fewer
+#'   observations selected, it is obeyed.
 #'
 #' @param mapping The aesthetic mapping, usually constructed with
 #'   \code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. Only needs
@@ -29,6 +53,10 @@
 #' @param keep.sparse logical If \code{TRUE}, the default, observations from the
 #'   more sparse regions are retained, if \code{FALSE} those from the densest
 #'   regions.
+#' @param keep.these character vector, integer vector, logical vector or
+#'   function that takes the variable mapped to the \code{label} aesthetic as
+#'   first argument and returns a character vector or a logical vector. These
+#'   rows from \code{data} are selected irrespective of the local density.
 #' @param invert.selection logical If \code{TRUE}, the complement of the
 #'   selected rows are returned.
 #' @param h vector of bandwidths for x and y directions. Defaults to normal
@@ -36,7 +64,7 @@
 #'   apply to both directions.
 #' @param n Number of grid points in each direction. Can be scalar or a length-2
 #'   integer vector
-#' @param label.fill character vector of length 1 or a function.
+#' @param label.fill character vector of length 1, a function or \code{NULL}.
 #' @param position The position adjustment to use for overlapping points on this
 #'   layer
 #' @param show.legend logical. Should this layer be included in the legends?
@@ -55,7 +83,10 @@
 #' @return A plot layer instance. Using as output \code{data} the input
 #'   \code{data} after value substitution based on a 2D the filtering criterion.
 #'
-#' @seealso \code{\link[MASS]{kde2d}} used internally.
+#' @seealso \code{\link{stat_dens2d_filter}} and \code{\link[MASS]{kde2d}} used
+#'   internally. Parameters \code{n}, \code{h} in this statistic correspond to
+#'   the parameters with the same name in this imported function. Limits are set
+#'   to the limits of the plot scales.
 #'
 #' @family statistics returning a subset of data
 #'
@@ -81,6 +112,14 @@
 #' ggplot(data = d, aes(x, y, label = lab)) +
 #'   geom_point() +
 #'   stat_dens2d_labels()
+#'
+#' ggplot(data = d, aes(x, y, label = lab)) +
+#'   geom_point() +
+#'   stat_dens2d_labels(keep.these = "zoujdg")
+#'
+#' ggplot(data = d, aes(x, y, label = lab)) +
+#'   geom_point() +
+#'   stat_dens2d_labels(keep.these = function(x) {grepl("^z", x)})
 #'
 #' ggplot(data = d, aes(x, y, label = lab)) +
 #'   geom_point() +
@@ -122,6 +161,10 @@
 #'   ggplot(data = d, aes(x, y, label = lab)) +
 #'     geom_point() +
 #'     stat_dens2d_labels(geom = "debug")
+#'
+#'   ggplot(data = d, aes(x, y)) +
+#'     geom_point() +
+#'     stat_dens2d_labels(geom = "debug")
 #' }
 #'
 stat_dens2d_labels <-
@@ -133,6 +176,7 @@ stat_dens2d_labels <-
            keep.fraction = 0.10,
            keep.number = Inf,
            keep.sparse = TRUE,
+           keep.these = FALSE,
            invert.selection = FALSE,
            h = NULL,
            n = NULL,
@@ -140,6 +184,17 @@ stat_dens2d_labels <-
            na.rm = TRUE,
            show.legend = FALSE,
            inherit.aes = TRUE) {
+
+    if (length(label.fill) > 1L) {
+      stop("Length for 'label.fill' is not 1: ", label.fill)
+    }
+    if (is.na(keep.fraction) || keep.fraction < 0 || keep.fraction > 1) {
+      stop("Out of range or missing value for 'keep.fraction': ", keep.fraction)
+    }
+    if (is.na(keep.number) || keep.number < 0) {
+      stop("Out of range or missing value for 'keep.number': ", keep.number)
+    }
+
     ggplot2::layer(
       stat = StatDens2dLabels, data = data, mapping = mapping, geom = geom,
       position = position, show.legend = show.legend, inherit.aes = inherit.aes,
@@ -147,6 +202,7 @@ stat_dens2d_labels <-
                     keep.fraction = keep.fraction,
                     keep.number = keep.number,
                     keep.sparse = keep.sparse,
+                    keep.these = keep.these,
                     invert.selection = invert.selection,
                     h = h,
                     n = n,
@@ -161,30 +217,22 @@ dens2d_labs_compute_fun <-
            keep.fraction,
            keep.number,
            keep.sparse,
+           keep.these,
            invert.selection,
            h,
            n,
            label.fill) {
 
-    if (length(label.fill) != 1L) {
-      stop("Length for 'label.fill' is not 1: ", label.fill)
-    }
-
-    if (is.na(keep.fraction) || keep.fraction < 0 || keep.fraction > 1) {
-      stop("Out of range or missing value for 'keep.fraction': ", keep.fraction)
-    }
-    if (is.na(keep.number) || keep.number < 0) {
-      stop("Out of range or missing value for 'keep.number': ", keep.number)
-    }
-
     force(data)
+    if (!exists("label", data)) {
+#      message("Mapping 'rownames(data)' to missing 'label' aesthetic")
+      data[["label"]] <- rownames(data)
+    }
+
+    keep.these <- keep_these2logical(keep.these = keep.these, data = data)
+
     if (nrow(data) * keep.fraction > keep.number) {
       keep.fraction <- keep.number / nrow(data)
-    }
-
-    if (!exists("label", data)) {
-      message("Mapping missing 'label' aesthetic to 'rownames(data)'.")
-      data[["label"]] <- rownames(data)
     }
 
     if (keep.fraction == 1) {
@@ -219,13 +267,19 @@ dens2d_labs_compute_fun <-
       }
     }
 
-    if (is.function(label.fill)) {
+    if (is.null(label.fill)) {
       if (invert.selection){
-        data[["label"]] <- ifelse(!keep,
+        data <- data[!(keep | keep.these), ]
+      } else {
+        data <- data[keep | keep.these, ]
+      }
+    } else if (is.function(label.fill)) {
+      if (invert.selection){
+        data[["label"]] <- ifelse(!(keep | keep.these),
                                   data[["label"]],
                                   label.fill(data[["label"]]))
       } else {
-        data[["label"]] <- ifelse(keep,
+        data[["label"]] <- ifelse(keep | keep.these,
                                   data[["label"]],
                                   label.fill(data[["label"]]))
       }
@@ -240,11 +294,11 @@ dens2d_labs_compute_fun <-
         }
       }
       if (invert.selection){
-        data[["label"]] <- ifelse(!keep,
+        data[["label"]] <- ifelse(!(keep | keep.these),
                                   data[["label"]],
                                   label.fill)
       } else {
-        data[["label"]] <- ifelse(keep,
+        data[["label"]] <- ifelse(keep | keep.these,
                                   data[["label"]],
                                   label.fill)
       }
@@ -265,3 +319,41 @@ StatDens2dLabels <-
       dens2d_labs_compute_fun,
     required_aes = c("x", "y")
   )
+
+
+# Utils for stats that subset data
+
+#' Convert keep.these argument into logical vector
+#'
+#' @param keep.these character vector, integer vector, logical vector or
+#'   function that takes the variable mapped to the \code{label} aesthetic as
+#'   first argument and returns a character vector or a logical vector. These
+#'   rows from \code{data} are selected irrespective of the local density.
+#' @param data data.frame The plot layer's data set.
+#'
+#' @keywords internal
+#'
+keep_these2logical <- function(keep.these, data) {
+  if (length(keep.these)) {
+    if (is.function(keep.these)) {
+      keep.these <- keep.these(data$label) # character or logical vector
+    }
+    if (is.character(keep.these)) {
+      keep.these <- data$label %in% keep.these # logical vector
+    }
+    if (is.numeric(keep.these)) {
+      temp <- logical(nrow(data))
+      temp[keep.these] <- TRUE
+      keep.these <- temp
+    }
+    if (anyNA(keep.these)) {
+      warning("Discarding 'NA's in keep.these")
+      keep.these <- ifelse(is.na(keep.these),
+                           FALSE,
+                           keep.these)
+    }
+  } else { # replace NULL and vectors with length zero with FALSE
+    keep.these <- FALSE
+  }
+  keep.these
+}